backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
deprecation==2.1.0
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.57.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.3
mindee==3.13.1
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pikepdf==6.2.9
Pillow==10.0.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
